Zip certain TM1 files
Rather often I want to backup only processes or rules files (or both) - just as a quick safety measure or to transfer (only) these files out of a TM1 data directory. This PowerShell script  can do this for us. Drop it where you need it, change the folder names near the top of the script and you are good to go.